% Auto-generated: do not edit by hand
\name{''DashFlow}

\alias{''DashFlow}

\title{DashFlow component}

\description{

}

\usage{
''DashFlow(children=NULL, id=NULL, clickedEdge=NULL, clickedNode=NULL,
connectedEdge=NULL, dagre_direction=NULL, edges=NULL,
nodes=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{clickedEdge}{Named list. }

\item{clickedNode}{Named list. }

\item{connectedEdge}{Named list. }

\item{dagre_direction}{Character. }

\item{edges}{Unnamed list. }

\item{nodes}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

