% Auto-generated: do not edit by hand
\name{''MiniMap}

\alias{''MiniMap}

\title{MiniMap component}

\description{

}

\usage{
''MiniMap(id=NULL, className=NULL, maskColor=NULL,
nodeBorderRadius=NULL, nodeClassName=NULL, nodeColor=NULL,
nodeStrokeWidth=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{maskColor}{Character. }

\item{nodeBorderRadius}{Numeric. }

\item{nodeClassName}{Character. }

\item{nodeColor}{Character. }

\item{nodeStrokeWidth}{Numeric. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

